pyproject.toml
[project]
name = "parakeet"
version = "0.1.0"
description = "Windows tray transcription tray app foundation"
authors = [{ name = "Parakeet Contributors", email = "maintainers@parakeet.app" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<4.0"
dependencies = [
    "Pillow>=10.4.0",
    "pystray>=0.19.5",
    "pynput>=1.7.7",
    "sounddevice (>=0.5.3,<0.6.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "webrtcvad (>=2.0.10,<3.0.0)",
    "onnxruntime-directml (>=1.20.0,<2.0.0)",
    "onnx-asr (>=0.1.0)",
    "pyperclip (>=1.8.2,<2.0.0)",
    "huggingface-hub (>=0.25.0,<1.0.0)",
]

license-files = ["LICENSE"]

[tool.poetry]
packages = [{ include = "app", from = "src" }]
include = ["src/app/py.typed"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
mypy = "^1.11.2"
pre-commit = "^3.8.0"
pytest = "^8.3.3"
pyinstaller = { version = "^6.10.0", python = "<3.14" }

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "B", "I"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
packages = ["app"]
ignore_missing_imports = true
show_error_codes = true
pretty = true

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
